[#14]refactor: services -> service
