+++ app/route/execute/execute_router.py
@@ -3,17 +3,17 @@ from starlette.responses import JSONResponse
 
 from app.web.models.success_reponse import SuccessResponse
 from app.route.execute.models.code_request import CodeRequest
-from app.route.execute.services import analsys_service, execute_service, syntax_service
+from app.route.execute.service import analsys_service, execute_service, syntax_service
 
 execute_router = APIRouter()
 
 
 @execute_router.post("/v1/execute/visualize")
-async def visualize(code: CodeRequest):
+async def visualize(code_request: CodeRequest):
     # 문법 체크
-    code = syntax_service.check(code.source_code)
+    execute_service.execute_code(code_request.source_code, code_request.input)
     # 코드 분석
-    analysis_result = analsys_service.analyze_code(code)
+    analysis_result = analsys_service.analyze_code(code_request.source_code, code_request.input)
 
     success_response = SuccessResponse(
         detail="success code analysis",
